0.17.1.5-Mandatory Upgrade for Entire Network @320,000
- Fix arm64 bug (allows processor floating point rounding errors for
deflation calculation).  This was in a prior version of biblepay, but
though we had grown out of it so it was left out, but it needs in.
- This above bug can cause you not to sync on a warm node no matter
which architecture.  Simply upgrade and restart the client.  If this
does not work use -erasechain=1 once.  Going forward it will be resolved
automatically without resyncing.
- Kick off old versions prior to mandatory upgrade to give us a new
network (prior to peer_pro_version 70777).