0.21.5-Mandatory Upgrade
Summary of Changes for 0.21.5 - Hamans Hanging:

- Added NFT add,edit,delete capability to core wallet.
- Reinstituted SideChain transactions.
- Incremented min_masternode_proto_version to 70783 so as to facilitate a mandatory upgrade for LLMQ messages.
(Requires Temples to enforce our global LLMQ).
- Added CheckMemPoolTransactionBiblepay, which is our facility
to check mempool txs.  In this version we enforce the NFT buy rules, to allow the safe transfer of NFT.
- Reinstituted embedded sTxOutMessage in CreateRawTransaction (allows longer messages, such as NFTs to go in the chain).
- Prevent Sancs from continually adding false governance triggers due to a bug.
The code checks for an existing govobj first, and also caches the last one created (preventing dupes).
- Ensure sancs do not vote on triggers that already won.
- Make temples handle WatchmanOnTheWall, but not Sancs.
- Added exec listnfts.
- Removed UI deadlock (affecting various RPC commands
like sendrawtransaction and upgradesanc, and possibly the LLMQ
timeouts).
- Modified the LLMQ rules to work with Temples.
- Modified Pose background thread to only deal with one
sanctuary at a time (solves the deadlock).